move test_topologyattrs to right place
